This is a bit specific to my input.
I was going to try doing something cleverer with breadth first searching and leaving breadcrumbs
for predecessor calculations, but figuring out when I came across parts of the path I'd seen
before but had found a better path to got too messy - how do I notify the other path that it
has a better route to it now? It might even have given up spuriously already as it exceeded
some path length to the end. Much easier to crank up the exploration limit and be done using
networkx have it handle the hard work.